requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
algoliasearch==3.0.0
openai==1.3.8
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
beautifulsoup4==4.12.2
feedgen==1.0.0
gunicorn==21.2.0
